requirements.txt
requests==2.31.0
PyGithub==2.1.1
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
markdown==3.5
apscheduler==3.10.4
gunicorn==21.2.0
pymongo==4.6.1
pandas==2.1.3 